+++ src/extension.ts
@@ -1,9 +1,11 @@
 import * as vscode from 'vscode';
 
 export function activate(context: vscode.ExtensionContext) {
-	let editor = vscode.window.activeTextEditor;
+	let editor : vscode.TextEditor | undefined = vscode.window.activeTextEditor;
+	let shortcuts : vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration('fasttyping.shortcuts');
 	let lastTwoChars : string = '';
 	let countSelection : number = 0;
+	let fileExtension : string | undefined = editor?.document.fileName.split('.').pop();
 
 	let disposable = vscode.workspace.onDidChangeTextDocument(event => {
 		editor = vscode.window.activeTextEditor;
@@ -19,27 +21,50 @@ export function activate(context: vscode.ExtensionContext) {
 				lastTwoChars = '';
 				return;
 			}
+
 			if (lastTwoChars.length == 1) {
 				lastTwoChars = lastTwoChars + change.text;
 				// console.log("[1] DEBUG : " + lastTwoChars);
-				if (lastTwoChars === '--') {
-					// console.log("[1] DEBUG START : " + change.range.start.translate(0, -1).character);
-					// console.log("[1] DEBUG END : " + change.range.end.translate(0, 1 - countSelection).character);
-					const replaceRange = new vscode.Range(change.range.start.translate(0, -1), change.range.end.translate(0, 1 - countSelection));
-					editor?.edit(edit => edit.replace(replaceRange, '->'));
-				}
+	
+				Object.keys(shortcuts).forEach(key => {
+					let shortcut = shortcuts[key][lastTwoChars];
+					if (shortcut) {
+						// console.log(`Shortcut exists for ${key}: ${shortcut}`);
+						
+						// On a un raccourci pour cette extension de fichier
+						// Ou c'est un racourci global (et le raccourci n'existe pas pour le fichier)
+						// On vérifie aussi si l'extension est dans les clés de raccourcis sinon on peut pas
+						// vérifier si le raccourci pour l'extension existe
+						if (
+							key == fileExtension || 
+							(
+								key == 'common' &&
+									(
+										!shortcuts.hasOwnProperty(fileExtension!) ||
+										!JSON.stringify(shortcuts[fileExtension!]).includes(shortcut)
+									)
+							)
+						)  {
+							// console.log("[1] DEBUG START : " + change.range.start.translate(0, -1).character);
+							// console.log("[1] DEBUG END : " + change.range.end.translate(0, 1 - countSelection).character);
+							const replaceRange = new vscode.Range(change.range.start.translate(0, -1), change.range.end.translate(0, 1 - countSelection));
+							editor?.edit(edit => edit.replace(replaceRange, shortcut));
+						}
+					}
+				});
 			}
         }
     });
 
 	let subOnDidChangeTextEditorSelection = vscode.window.onDidChangeTextEditorSelection(async function (event) {
-		let editor = event.textEditor;
+		editor = event.textEditor;
 		if (!editor) {
 			return;
 		}
 
 		// On recupere la selection
-		let selection = editor.selection;
+		let selection : vscode.Selection = editor.selection;
+		fileExtension = editor.document.fileName.split('.').pop();
 
 		countSelection = 0;
 		// On a une selection donc on recupere le nombre de caractere selectionné
@@ -68,8 +93,28 @@ export function activate(context: vscode.ExtensionContext) {
 		// console.log("[MOVED] DEBUG : " + charBeforeSelection);
 	});
 
+	let subOnDidChangeConfiguration = vscode.workspace.onDidChangeConfiguration(event => {
+		if (event.affectsConfiguration('fasttyping.shortcuts')) {
+		  	// Les raccourcis ont été modifiés
+		  	shortcuts = vscode.workspace.getConfiguration('fasttyping.shortcuts');
+		}
+	});
+
+	// Enregistrer la commande pour ouvrir la configuration de l'extension
+	context.subscriptions.push(
+        vscode.commands.registerCommand('fasttyping.openConfig', openConfig)
+    );
+
     context.subscriptions.push(disposable);
 	context.subscriptions.push(subOnDidChangeTextEditorSelection);
+	context.subscriptions.push(subOnDidChangeConfiguration);
+}
+
+/**
+ * Ouvre la configuration de l'extension dans les paramètres utilisateur
+ */
+function openConfig() {
+	vscode.commands.executeCommand('workbench.action.openSettingsJson', '@ext:fasttyping');
 }
 
 export function deactivate() {}